time to BigInteger

+++ src/model/Point.java
@@ -1,17 +1,19 @@
 package model;
 
+import java.math.BigInteger;
+
 /**
  *
  * @author bram
  */
 public class Point {
 
-	public int t;
+	public BigInteger t;
 	public double x;
 	public double y;
 	public double z;
 
-	public Point(int t, double x, double y, double z) {
+	public Point(BigInteger t, double x, double y, double z) {
 		this.t = t;
 		this.x = x;
 		this.y = y;
